requirements.txt
PyQt5
Pillow
pytesseract
openai
python-dotenv
requests
pyinstaller
